+++ api/v2/api_pb/http.pb.gw.go
@@ -2,11 +2,11 @@
 // source: http.proto
 
 /*
-Package pb is a reverse proxy.
+Package api_pb is a reverse proxy.
 
 It translates gRPC into RESTful JSON APIs.
 */
-package pb
+package api_pb
 
 import (
 	"context"
