+++ api/v2/generate.sh
